[fix] Separates bcrypt hashing functions in separate file to resolve circular import
